updates to tf2.x
